+++ libretro.cpp
@@ -1,15 +1,16 @@
 #include <stdarg.h>
 #include <math.h>
-#include "mednafen/mednafen.h"
-#include "mednafen/git.h"
-#include "mednafen/general.h"
-#include "libretro.h"
 
+#include <libretro.h>
 #include <retro_timers.h>
 #include <streams/file_stream.h>
 
+#include "mednafen/mednafen.h"
+#include "mednafen/git.h"
+#include "mednafen/general.h"
 #include "mednafen/FileWrapper.h"
 
+#include "mednafen/pce_fast/huc.h"
 #include "mednafen/pce_fast/pce.h"
 #include "mednafen/pce_fast/vdc.h"
 #include "mednafen/pce_fast/psg.h"
@@ -28,36 +29,11 @@
 
 #include "libretro_core_options.h"
 
-#define MEDNAFEN_CORE_NAME_MODULE "pce_fast"
-#define MEDNAFEN_CORE_NAME "Mednafen SuperGrafx"
-#define MEDNAFEN_CORE_VERSION "v0.9.41"
-#define MEDNAFEN_CORE_EXTENSIONS "pce|sgx|cue|ccd|chd"
-#define MEDNAFEN_CORE_TIMING_FPS 59.82
-#define MEDNAFEN_CORE_GEOMETRY_BASE_W 512
-#define MEDNAFEN_CORE_GEOMETRY_BASE_H 243
-#define MEDNAFEN_CORE_GEOMETRY_MAX_W 512
-#define MEDNAFEN_CORE_GEOMETRY_MAX_H 243
-#define MEDNAFEN_CORE_GEOMETRY_ASPECT_RATIO (6.0 / 5.0)
-#define FB_WIDTH 512
-#define FB_HEIGHT 243
-
-#ifdef _WIN32
-static char slash = '\\';
-#else
-static char slash = '/';
-#endif
-
 static bool old_cdimagecache = false;
-static std::string retro_base_directory;
+std::string retro_base_directory;
 
 extern MDFNGI EmulatedPCE_Fast;
 MDFNGI *MDFNGameInfo = &EmulatedPCE_Fast;
-
-static int HuCLoad(const uint8 *data, uint32 len, uint32 crc32) MDFN_COLD;
-static int HuCLoadCD(const char *bios_path) MDFN_COLD;
-static void HuC_Close(void) MDFN_COLD;
-static int HuC_StateAction(StateMem *sm, int load, int data_only);
-static void HuC_Power(void) MDFN_COLD;
 
 /* Mednafen - Multi-system Emulator
  *
@@ -75,853 +51,6 @@ static void HuC_Power(void) MDFN_COLD;
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
-
-extern "C" unsigned long crc32(unsigned long crc, const unsigned char *buf, unsigned int len);
-
-static PCEFast_PSG *psg = NULL;
-extern ArcadeCard *arcade_card; // Bah, lousy globals.
-
-static Blip_Buffer sbuf[2];
-
-bool PCE_ACEnabled;
-
-static bool IsSGX;
-int pce_overclocked;
-
-// Statically allocated for speed...or something.
-uint8 ROMSpace[0x88 * 8192 + 8192];	// + 8192 for PC-as-pointer safety padding
-
-uint8 BaseRAM[32768 + 8192]; // 8KB for PCE, 32KB for Super Grafx // + 8192 for PC-as-pointer safety padding
-
-uint8 PCEIODataBuffer;
-
-static DECLFR(PCEBusRead)
-{
- //printf("BUS Read: %02x %04x\n", A >> 13, A);
- return(0xFF);
-}
-
-static DECLFW(PCENullWrite)
-{
- //printf("Null Write: %02x, %08x %02x\n", A >> 13, A, V);
-}
-
-static DECLFR(BaseRAMReadSGX)
-{
- return BaseRAM[(size_t)A - (0xF8 * 8192)];
-}
-
-static DECLFW(BaseRAMWriteSGX)
-{
- BaseRAM[(size_t)A - (0xF8 * 8192)] = V;
-}
-
-static DECLFR(BaseRAMRead)
-{
- return BaseRAM[(size_t)A - (0xF8 * 8192)];
-}
-
-static DECLFR(BaseRAMRead_Mirrored)
-{
- return(BaseRAM[A & 0x1FFF]);
-}
-
-static DECLFW(BaseRAMWrite)
-{
- BaseRAM[(size_t)A - (0xF8 * 8192)] = V;
-}
-
-static DECLFW(BaseRAMWrite_Mirrored)
-{
- BaseRAM[A & 0x1FFF] = V;
-}
-
-static DECLFR(IORead)
-{
- #define IOREAD_SGX 0
- #include "mednafen/pce_fast/ioread.inc"
- #undef IOREAD_SGX
-}
-
-static DECLFR(IOReadSGX)
-{
- #define IOREAD_SGX 1
- #include "mednafen/pce_fast/ioread.inc"
- #undef IOREAD_SGX
-}
-
-static DECLFW(IOWrite)
-{
- A &= 0x1FFF;
-
- switch(A >> 10)
- {
-  case 0: HuC6280_StealCycle();
-	       VDC_Write(A, V);
-	       break;
-
-  case 1: HuC6280_StealCycle();
-	       VCE_Write(A, V);
-	       break;
-
-  case 2: PCEIODataBuffer = V;
-	       psg->Write(HuCPU.timestamp / pce_overclocked, A, V);
-	       break;
-
-  case 3: PCEIODataBuffer = V;
-	       HuC6280_TimerWrite(A, V);
-	       break;
-
-  case 4: PCEIODataBuffer = V; INPUT_Write(A, V); break;
-  case 5: PCEIODataBuffer = V; HuC6280_IRQStatusWrite(A, V); break;
-  case 6:
-	  if(!PCE_IsCD)
-	   break;
-
-	  if((A & 0x1E00) == 0x1A00)
-	  {
-	   if(arcade_card)
-	    arcade_card->Write(A & 0x1FFF, V);
-	  }
-	  else
-	  {
-	   PCECD_Write(HuCPU.timestamp * 3, A, V);
-	  }
-	  break;
-
-  case 7: break;	// Expansion.
- }
-}
-
-static void PCECDIRQCB(bool asserted)
-{
- if(asserted)
-  HuC6280_IRQBegin(MDFN_IQIRQ2);
- else
-  HuC6280_IRQEnd(MDFN_IQIRQ2);
-}
-
-bool PCE_InitCD(void)
-{
- PCECD_Settings cd_settings;
- memset(&cd_settings, 0, sizeof(PCECD_Settings));
-
- cd_settings.CDDA_Volume = (double)MDFN_GetSettingUI("pce_fast.cddavolume") / 100;
- cd_settings.CD_Speed = MDFN_GetSettingUI("pce_fast.cdspeed");
-
- cd_settings.ADPCM_Volume = (double)MDFN_GetSettingUI("pce_fast.adpcmvolume") / 100;
- cd_settings.ADPCM_LPF = MDFN_GetSettingB("pce_fast.adpcmlp");
-
- if(cd_settings.CDDA_Volume != 1.0)
-  MDFN_printf(_("CD-DA Volume: %d%%\n"), (int)(100 * cd_settings.CDDA_Volume));
-
- if(cd_settings.ADPCM_Volume != 1.0)
-  MDFN_printf(_("ADPCM Volume: %d%%\n"), (int)(100 * cd_settings.ADPCM_Volume));
-
- return(PCECD_Init(&cd_settings, PCECDIRQCB, PCE_MASTER_CLOCK, pce_overclocked, &sbuf[0], &sbuf[1]));
-}
-
-static int LoadCommon(void);
-static void LoadCommonPre(void);
-
-static int Load(const char *name, MDFNFILE *fp)
-{
-   uint32 headerlen = 0;
-   uint32 r_size;
-
-   IsSGX = 0;
-
-   LoadCommonPre();
-
-   {
-      if(fp->size & 0x200) // 512 byte header!
-         headerlen = 512;
-   }
-
-   r_size = fp->size - headerlen;
-   if(r_size > 4096 * 1024)
-      r_size = 4096 * 1024;
-
-   for(int x = 0; x < 0x100; x++)
-   {
-      HuCPU.PCERead[x] = PCEBusRead;
-      HuCPU.PCEWrite[x] = PCENullWrite;
-   }
-
-   uint32 crc = crc32(0, fp->data + headerlen, fp->size - headerlen);
-
-   HuCLoad(fp->data + headerlen, fp->size - headerlen, crc);
-
-   if(!strcasecmp(fp->ext, "sgx"))
-      IsSGX = TRUE;
-
-   if(fp->size >= 8192 && !memcmp(fp->data + headerlen, "DARIUS Version 1.11b", strlen("DARIUS VERSION 1.11b")))
-   {
-      MDFN_printf("SuperGrafx:  Darius Plus\n");
-      IsSGX = 1;
-   }
-
-   if(crc == 0x4c2126b0)
-   {
-      MDFN_printf("SuperGrafx:  Aldynes\n");
-      IsSGX = 1;
-   }
-
-   if(crc == 0x8c4588e2)
-   {
-      MDFN_printf("SuperGrafx:  1941 - Counter Attack\n");
-      IsSGX = 1;
-   }
-   if(crc == 0x1f041166)
-   {
-      MDFN_printf("SuperGrafx:  Madouou Granzort\n");
-      IsSGX = 1;
-   }
-   if(crc == 0xb486a8ed)
-   {
-      MDFN_printf("SuperGrafx:  Daimakaimura\n");
-      IsSGX = 1;
-   }
-   if(crc == 0x3b13af61)
-   {
-      MDFN_printf("SuperGrafx:  Battle Ace\n");
-      IsSGX = 1;
-   }
-
-   // Space Harrier (Japan)/(USA) is not compatible with SuperGrafx mode
-   if (crc == 0x64580427 || crc == 0x43b05eb8)
-      IsSGX = 0;
-
-   if(crc == 0xfae0fc60)
-      OrderOfGriffonFix = true;
-
-   return(LoadCommon());
-}
-
-static void LoadCommonPre(void)
-{
- HuC6280_Init();
-
- // FIXME:  Make these globals less global!
- pce_overclocked = MDFN_GetSettingUI("pce_fast.ocmultiplier");
- PCE_ACEnabled = MDFN_GetSettingB("pce_fast.arcadecard");
-
- if(pce_overclocked > 1)
-  MDFN_printf(_("CPU overclock: %dx\n"), pce_overclocked);
-
- if(MDFN_GetSettingUI("pce_fast.cdspeed") > 1)
-  MDFN_printf(_("CD-ROM speed:  %ux\n"), (unsigned int)MDFN_GetSettingUI("pce_fast.cdspeed"));
-
- for(int x = 0; x < 0x100; x++)
- {
-  HuCPU.PCERead[x] = PCEBusRead;
-  HuCPU.PCEWrite[x] = PCENullWrite;
- }
-
- MDFNMP_Init(1024, (1 << 21) / 1024);
-}
-
-static int LoadCommon(void)
-{
- IsSGX |= MDFN_GetSettingB("pce_fast.forcesgx") ? 1 : 0;
-
- // Don't modify IsSGX past this point.
-
- VDC_Init(IsSGX);
-
- if(IsSGX)
- {
-  MDFN_printf("SuperGrafx Emulation Enabled.\n");
-  HuCPU.PCERead[0xF8] = HuCPU.PCERead[0xF9] = HuCPU.PCERead[0xFA] = HuCPU.PCERead[0xFB] = BaseRAMReadSGX;
-  HuCPU.PCEWrite[0xF8] = HuCPU.PCEWrite[0xF9] = HuCPU.PCEWrite[0xFA] = HuCPU.PCEWrite[0xFB] = BaseRAMWriteSGX;
-
-  for(int x = 0xf8; x < 0xfb; x++)
-   HuCPU.FastMap[x] = &BaseRAM[(x & 0x3) * 8192];
-
-  HuCPU.PCERead[0xFF] = IOReadSGX;
- }
- else
- {
-  HuCPU.PCERead[0xF8] = BaseRAMRead;
-  HuCPU.PCERead[0xF9] = HuCPU.PCERead[0xFA] = HuCPU.PCERead[0xFB] = BaseRAMRead_Mirrored;
-
-  HuCPU.PCEWrite[0xF8] = BaseRAMWrite;
-  HuCPU.PCEWrite[0xF9] = HuCPU.PCEWrite[0xFA] = HuCPU.PCEWrite[0xFB] = BaseRAMWrite_Mirrored;
-
-  for(int x = 0xf8; x < 0xfb; x++)
-   HuCPU.FastMap[x] = &BaseRAM[0];
-
-  HuCPU.PCERead[0xFF] = IORead;
- }
-
- MDFNMP_AddRAM(IsSGX ? 32768 : 8192, 0xf8 * 8192, BaseRAM);
-
- HuCPU.PCEWrite[0xFF] = IOWrite;
-
- psg = new PCEFast_PSG(&sbuf[0], &sbuf[1]);
-
- psg->SetVolume(1.0);
-
- if(PCE_IsCD)
- {
-  unsigned int cdpsgvolume = MDFN_GetSettingUI("pce_fast.cdpsgvolume");
-
-  if(cdpsgvolume != 100)
-  {
-   MDFN_printf(_("CD PSG Volume: %d%%\n"), cdpsgvolume);
-  }
-
-  psg->SetVolume(0.678 * cdpsgvolume / 100);
-
- }
-
- PCEINPUT_Init();
-
- PCE_Power();
-
-#if 0
- MDFNGameInfo->LayerNames = IsSGX ? "BG0\0SPR0\0BG1\0SPR1\0" : "Background\0Sprites\0";
-#endif
- MDFNGameInfo->fps = (uint32)((double)7159090.90909090 / 455 / 263 * 65536 * 256);
-
- return(1);
-}
-
-#ifdef _WIN32
-static void sanitize_path(std::string &path)
-{
-   size_t size = path.size();
-   for (size_t i = 0; i < size; i++)
-      if (path[i] == '/')
-         path[i] = '\\';
-}
-#endif
-
-static int LoadCD(std::vector<CDIF *> *CDInterfaces)
-{
- std::string bios_path = retro_base_directory + slash + MDFN_GetSettingS("pce_fast.cdbios");
-
-#ifdef _WIN32
- sanitize_path(bios_path);
-#endif
-
- if (log_cb)
-  log_cb(RETRO_LOG_INFO, "Loading bios %s\n", bios_path.c_str());
-
- IsSGX = 0;
-
- LoadCommonPre();
-
- if(!HuCLoadCD(bios_path.c_str()))
-  return(0);
-
- PCECD_Drive_SetDisc(true, NULL, true);
- PCECD_Drive_SetDisc(false, (*CDInterfaces)[0], true);
-
- return(LoadCommon());
-}
-
-static void Cleanup_PCE(void)
-{
-   HuC_Close();
-
-   VDC_Close();
-
-   if(psg)
-      delete psg;
-   psg = NULL;
-}
-
-static void CloseGame(void)
-{
-   Cleanup_PCE();
-}
-
-static void Emulate(EmulateSpecStruct *espec)
-{
- INPUT_Frame();
-
- MDFNMP_ApplyPeriodicCheats();
-
- #if 0
- {
-  static bool firstcat = true;
-  MDFN_PixelFormat nf;
-
-  nf.bpp = 16;
-  nf.colorspace = MDFN_COLORSPACE_RGB;
-  nf.Rshift = 11;
-  nf.Gshift = 5;
-  nf.Bshift = 0;
-  nf.Ashift = 16;
-
-  nf.Rprec = 5;
-  nf.Gprec = 6;
-  nf.Bprec = 5;
-  nf.Aprec = 8;
-
-  espec->surface->SetFormat(nf, false);
-  espec->VideoFormatChanged = firstcat;
-  firstcat = false;
- }
- #endif
-
-#if 0
- static bool firstcat = true;
-
- MDFN_PixelFormat tmp_pf;
-
- tmp_pf.Rshift = 0;
- tmp_pf.Gshift = 0;
- tmp_pf.Bshift = 0;
- tmp_pf.Ashift = 8;
-
- tmp_pf.Rprec = 0;
- tmp_pf.Gprec = 0;
- tmp_pf.Bprec = 0;
- tmp_pf.Aprec = 0;
-
- tmp_pf.bpp = 8;
- tmp_pf.colorspace = MDFN_COLORSPACE_RGB;
-
- espec->surface->SetFormat(tmp_pf, false);
- espec->VideoFormatChanged = firstcat;
- firstcat = false;
-#endif
-
- if(espec->VideoFormatChanged)
-  VDC_SetPixelFormat(espec->surface->format); //.Rshift, espec->surface->format.Gshift, espec->surface->format.Bshift);
-
- if(espec->SoundFormatChanged)
- {
-  for(int y = 0; y < 2; y++)
-  {
-   sbuf[y].set_sample_rate(espec->SoundRate ? espec->SoundRate : 44100, 50);
-   sbuf[y].clock_rate((long)(PCE_MASTER_CLOCK / 3));
-   sbuf[y].bass_freq(10);
-  }
- }
- VDC_RunFrame(espec, false);
-
- if(PCE_IsCD)
- {
-  PCECD_Run(HuCPU.timestamp * 3);
- }
-
- psg->EndFrame(HuCPU.timestamp / pce_overclocked);
-
- if(espec->SoundBuf)
- {
-  for(int y = 0; y < 2; y++)
-  {
-   sbuf[y].end_frame(HuCPU.timestamp / pce_overclocked);
-   espec->SoundBufSize = sbuf[y].read_samples(espec->SoundBuf + y, espec->SoundBufMaxSize, 1);
-  }
- }
-
- espec->MasterCycles = HuCPU.timestamp * 3;
-
- INPUT_FixTS();
-
- HuC6280_ResetTS();
-
- if(PCE_IsCD)
-  PCECD_ResetTS();
-}
-
-int StateAction(void *data, int load, int data_only)
-{
- SFORMAT StateRegs[] =
- {
-  SFARRAY(BaseRAM, IsSGX? 32768 : 8192),
-  SFVAR(PCEIODataBuffer),
-  SFEND
- };
- StateMem *sm = (StateMem*)data;
-
- //for(int i = 8192; i < 32768; i++)
- // if(BaseRAM[i] != 0xFF)
- //  printf("%d %02x\n", i, BaseRAM[i]);
-
- int ret = MDFNSS_StateAction(sm, load, data_only, StateRegs, "MAIN", false);
-
- ret &= HuC6280_StateAction(sm, load, data_only);
- ret &= VDC_StateAction(sm, load, data_only);
- ret &= psg->StateAction(sm, load, data_only);
- ret &= INPUT_StateAction(sm, load, data_only);
- ret &= HuC_StateAction(sm, load, data_only);
-
- if(load)
- { }
-
- return(ret);
-}
-
-void PCE_Power(void)
-{
- memset(BaseRAM, 0x00, sizeof(BaseRAM));
-
- if(!IsSGX)
-  for(int i = 8192; i < 32768; i++)
-   BaseRAM[i] = 0xFF;
-
- PCEIODataBuffer = 0xFF;
-
- HuC6280_Power();
- VDC_Power();
- psg->Power(HuCPU.timestamp / pce_overclocked);
- HuC_Power();
-
- if(PCE_IsCD)
- {
-  PCECD_Power(HuCPU.timestamp * 3);
- }
-}
-
-static void DoSimpleCommand(int cmd)
-{
- switch(cmd)
- {
-  case MDFN_MSC_RESET: PCE_Power(); break;
-  case MDFN_MSC_POWER: PCE_Power(); break;
- }
-}
-
-static MDFNSetting PCESettings[] =
-{
-  { "pce_fast.slstart", MDFNSF_NOFLAGS, "First rendered scanline.", NULL, MDFNST_UINT, "4", "0", "239" },
-  { "pce_fast.slend", MDFNSF_NOFLAGS, "Last rendered scanline.", NULL, MDFNST_UINT, "235", "0", "239" },
-  { "pce_fast.mouse_sensitivity", MDFNSF_NOFLAGS, "Mouse sensitivity.", NULL, MDFNST_FLOAT, "0.50", NULL, NULL, NULL, PCEINPUT_SettingChanged },
-  { "pce_fast.disable_softreset", MDFNSF_NOFLAGS, "If set, when RUN+SEL are pressed simultaneously, disable both buttons temporarily.", NULL, MDFNST_BOOL, "0", NULL, NULL, NULL, PCEINPUT_SettingChanged },
-  { "pce_fast.forcesgx", MDFNSF_EMU_STATE | MDFNSF_UNTRUSTED_SAFE, "Force SuperGrafx emulation.", NULL, MDFNST_BOOL, "0" },
-  { "pce_fast.arcadecard", MDFNSF_EMU_STATE | MDFNSF_UNTRUSTED_SAFE, "Enable Arcade Card emulation.", NULL, MDFNST_BOOL, "1" },
-  { "pce_fast.ocmultiplier", MDFNSF_EMU_STATE | MDFNSF_UNTRUSTED_SAFE, "CPU overclock multiplier.", NULL, MDFNST_UINT, "1", "1", "100"},
-  { "pce_fast.cdspeed", MDFNSF_EMU_STATE | MDFNSF_UNTRUSTED_SAFE, "CD-ROM data transfer speed multiplier", NULL, MDFNST_UINT, "1", "1", "100" },
-  { "pce_fast.nospritelimit", MDFNSF_NOFLAGS, "Remove 16-sprites-per-scanline hardware limit.", NULL, MDFNST_BOOL, "0" },
-  { "pce_fast.hoverscan", MDFNSF_NOFLAGS, "Reduce 352 pixels width overscan.", NULL, MDFNST_UINT, "300", "300", "352" },
-  { "pce_fast.cdbios", MDFNSF_EMU_STATE, "Path to the CD BIOS", NULL, MDFNST_STRING, "syscard3.pce" },
-  { "pce_fast.adpcmlp", MDFNSF_NOFLAGS, "Enable dynamic ADPCM lowpass filter.", NULL, MDFNST_BOOL, "0" },
-  { "pce_fast.cdpsgvolume", MDFNSF_NOFLAGS, "PSG volume when playing a CD game.", NULL, MDFNST_UINT, "100", "0", "200" },
-  { "pce_fast.cddavolume", MDFNSF_NOFLAGS, "CD-DA volume.", NULL, MDFNST_UINT, "100", "0", "200" },
-  { "pce_fast.adpcmvolume", MDFNSF_NOFLAGS, "ADPCM volume.", NULL, MDFNST_UINT, "100", "0", "200" },
-  { NULL }
-};
-
-uint8 MemRead(uint32 addr)
-{
- return(HuCPU.PCERead[(addr / 8192) & 0xFF](addr));
-}
-
-static const FileExtensionSpecStruct KnownExtensions[] =
-{
- { ".pce", "PC Engine ROM Image" },
- { ".hes", "PC Engine Music Rip" },
- { ".sgx", "SuperGrafx ROM Image" },
- { NULL, NULL }
-};
-
-static const uint8 BRAM_Init_String[8] = { 'H', 'U', 'B', 'M', 0x00, 0x88, 0x10, 0x80 }; //"HUBM\x00\x88\x10\x80";
-
-ArcadeCard *arcade_card = NULL;
-
-static uint8 *HuCROM = NULL;
-
-static bool IsPopulous;
-bool PCE_IsCD;
-
-uint8 SaveRAM[2048];
-
-static DECLFW(ACPhysWrite)
-{
- arcade_card->PhysWrite(A, V);
-}
-
-static DECLFR(ACPhysRead)
-{
- return(arcade_card->PhysRead(A));
-}
-
-static DECLFR(SaveRAMRead)
-{
- if((!PCE_IsCD || PCECD_IsBRAMEnabled()) && (A & 8191) < 2048)
-  return(SaveRAM[A & 2047]);
- else
-  return(0xFF);
-}
-
-static DECLFW(SaveRAMWrite)
-{
- if((!PCE_IsCD || PCECD_IsBRAMEnabled()) && (A & 8191) < 2048)
-  SaveRAM[A & 2047] = V;
-}
-
-static DECLFR(HuCRead)
-{
- return ROMSpace[A];
-}
-
-static DECLFW(HuCRAMWrite)
-{
- ROMSpace[A] = V;
-}
-
-static DECLFW(HuCRAMWriteCDSpecial) // Hyper Dyne Special hack
-{
- BaseRAM[0x2000 | (A & 0x1FFF)] = V;
- ROMSpace[A] = V;
-}
-
-static uint8 HuCSF2Latch = 0;
-
-static DECLFR(HuCSF2Read)
-{
- return(HuCROM[(A & 0x7FFFF) + 0x80000 + HuCSF2Latch * 0x80000 ]); // | (HuCSF2Latch << 19) ]);
-}
-
-static DECLFW(HuCSF2Write)
-{
- if((A & 0x1FFC) == 0x1FF0)
- {
-  HuCSF2Latch = (A & 0x3);
- }
-}
-
-static void Cleanup(void)
-{
-   if(arcade_card)
-      delete arcade_card;
-   arcade_card = NULL;
-
-   if(PCE_IsCD)
-      PCECD_Close();
-
-   if(HuCROM)
-      MDFN_free(HuCROM);
-   HuCROM = NULL;
-}
-
-static int HuCLoad(const uint8 *data, uint32 len, uint32 crc32)
-{
- uint32 sf2_threshold = 2048 * 1024;
- uint32 sf2_required_size = 2048 * 1024 + 512 * 1024;
- uint32 m_len = (len + 8191)&~8191;
- bool sf2_mapper = FALSE;
-
- if(m_len >= sf2_threshold)
- {
-  sf2_mapper = TRUE;
-
-  if(m_len != sf2_required_size)
-   m_len = sf2_required_size;
- }
-
- IsPopulous = 0;
- PCE_IsCD = 0;
-
- MDFN_printf(_("ROM:       %dKiB\n"), (len + 1023) / 1024);
- MDFN_printf(_("ROM CRC32: 0x%04x\n"), crc32);
-
- if(!(HuCROM = (uint8 *)MDFN_malloc(m_len, _("HuCard ROM"))))
-    return 0;
-
- memset(HuCROM, 0xFF, m_len);
- memcpy(HuCROM, data, (m_len < len) ? m_len : len);
-
- memset(ROMSpace, 0xFF, 0x88 * 8192 + 8192);
-
- if(m_len == 0x60000)
- {
-  memcpy(ROMSpace + 0x00 * 8192, HuCROM, 0x20 * 8192);
-  memcpy(ROMSpace + 0x20 * 8192, HuCROM, 0x20 * 8192);
-  memcpy(ROMSpace + 0x40 * 8192, HuCROM + 0x20 * 8192, 0x10 * 8192);
-  memcpy(ROMSpace + 0x50 * 8192, HuCROM + 0x20 * 8192, 0x10 * 8192);
-  memcpy(ROMSpace + 0x60 * 8192, HuCROM + 0x20 * 8192, 0x10 * 8192);
-  memcpy(ROMSpace + 0x70 * 8192, HuCROM + 0x20 * 8192, 0x10 * 8192);
- }
- else if(m_len == 0x80000)
- {
-  memcpy(ROMSpace + 0x00 * 8192, HuCROM, 0x40 * 8192);
-  memcpy(ROMSpace + 0x40 * 8192, HuCROM + 0x20 * 8192, 0x20 * 8192);
-  memcpy(ROMSpace + 0x60 * 8192, HuCROM + 0x20 * 8192, 0x20 * 8192);
- }
- else
- {
-  memcpy(ROMSpace + 0x00 * 8192, HuCROM, (m_len < 1024 * 1024) ? m_len : 1024 * 1024);
- }
-
- for(int x = 0x00; x < 0x80; x++)
- {
-  HuCPU.FastMap[x] = &ROMSpace[x * 8192];
-  HuCPU.PCERead[x] = HuCRead;
- }
-
- if(!memcmp(HuCROM + 0x1F26, "POPULOUS", strlen("POPULOUS")))
- {
-  uint8 *PopRAM = ROMSpace + 0x40 * 8192;
-
-  memset(PopRAM, 0xFF, 32768);
-
-  IsPopulous = 1;
-
-  MDFN_printf("Populous\n");
-
-  for(int x = 0x40; x < 0x44; x++)
-  {
-   HuCPU.FastMap[x] = &PopRAM[(x & 3) * 8192];
-   HuCPU.PCERead[x] = HuCRead;
-   HuCPU.PCEWrite[x] = HuCRAMWrite;
-  }
-  MDFNMP_AddRAM(32768, 0x40 * 8192, PopRAM);
- }
- else
- {
-  memset(SaveRAM, 0x00, 2048);
-  memcpy(SaveRAM, BRAM_Init_String, 8);    // So users don't have to manually intialize the file cabinet
-                                                // in the CD BIOS screen.
-  HuCPU.PCEWrite[0xF7] = SaveRAMWrite;
-  HuCPU.PCERead[0xF7] = SaveRAMRead;
-  MDFNMP_AddRAM(2048, 0xF7 * 8192, SaveRAM);
- }
-
- // 0x1A558
- //if(len >= 0x20000 && !memcmp(HuCROM + 0x1A558, "STREET FIGHTER#", strlen("STREET FIGHTER#")))
- if(sf2_mapper)
- {
-  for(int x = 0x40; x < 0x80; x++)
-  {
-   // FIXME: PCE_FAST
-   // HuCPUFastMap[x] = NULL; // Make sure our reads go through our read function, and not a table lookup
-   HuCPU.PCERead[x] = HuCSF2Read;
-  }
-  HuCPU.PCEWrite[0] = HuCSF2Write;
-  MDFN_printf("Street Fighter 2 Mapper\n");
-  HuCSF2Latch = 0;
- }
-
- return(1);
-}
-
-static int HuCLoadCD(const char *bios_path)
-{
-   MDFNFILE *fp = file_open(bios_path);
-
-   if (!fp)
-   {
-      if (log_cb)
-         log_cb(RETRO_LOG_ERROR, "Failed to load bios!\n");
-      return(0);
-  }
-
-   memset(ROMSpace, 0xFF, 262144);
-
-   memcpy(ROMSpace, fp->data + (fp->size & 512), ((fp->size & ~512) > 262144) ? 262144 : (fp->size &~ 512) );
-
-   file_close(fp);
-
-   PCE_IsCD = 1;
-   PCE_InitCD();
-
-   MDFN_printf(_("Arcade Card Emulation:  %s\n"), PCE_ACEnabled ? _("Enabled") : _("Disabled"));
-   for(int x = 0; x < 0x40; x++)
-   {
-      HuCPU.FastMap[x] = &ROMSpace[x * 8192];
-      HuCPU.PCERead[x] = HuCRead;
-   }
-
-   for(int x = 0x68; x < 0x88; x++)
-   {
-      HuCPU.FastMap[x] = &ROMSpace[x * 8192];
-      HuCPU.PCERead[x] = HuCRead;
-      HuCPU.PCEWrite[x] = HuCRAMWrite;
-   }
-   HuCPU.PCEWrite[0x80] = HuCRAMWriteCDSpecial; 	// Hyper Dyne Special hack
-   MDFNMP_AddRAM(262144, 0x68 * 8192, ROMSpace + 0x68 * 8192);
-
-   if(PCE_ACEnabled)
-   {
-      if (!(arcade_card = new ArcadeCard()))
-      {
-         MDFN_PrintError(_("Error creating %s object.\n"), "ArcadeCard");
-         Cleanup();
-         return(0);
-      }
-
-      for(int x = 0x40; x < 0x44; x++)
-      {
-         HuCPU.PCERead[x] = ACPhysRead;
-         HuCPU.PCEWrite[x] = ACPhysWrite;
-      }
-   }
-
-   memset(SaveRAM, 0x00, 2048);
-   memcpy(SaveRAM, BRAM_Init_String, 8);	// So users don't have to manually intialize the file cabinet
-   // in the CD BIOS screen.
-
-   HuCPU.PCEWrite[0xF7] = SaveRAMWrite;
-   HuCPU.PCERead[0xF7] = SaveRAMRead;
-   MDFNMP_AddRAM(2048, 0xF7 * 8192, SaveRAM);
-   return(1);
-}
-
-static int HuC_StateAction(StateMem *sm, int load, int data_only)
-{
- SFORMAT StateRegs[] =
- {
-  SFARRAY(ROMSpace + 0x40 * 8192, IsPopulous ? 32768 : 0),
-  SFARRAY(SaveRAM, IsPopulous ? 0 : 2048),
-  SFARRAY(ROMSpace + 0x68 * 8192, PCE_IsCD ? 262144 : 0),
-  SFVAR(HuCSF2Latch),
-  SFEND
- };
- int ret = MDFNSS_StateAction(sm, load, data_only, StateRegs, "HuC", false);
-
- if(load)
-  HuCSF2Latch &= 0x3;
-
- if(PCE_IsCD)
- {
-  ret &= PCECD_StateAction(sm, load, data_only);
-
-  if(arcade_card)
-   ret &= arcade_card->StateAction(sm, load, data_only);
- }
- return(ret);
-}
-
-static void HuC_Close(void)
-{
-   Cleanup();
-}
-
-static void HuC_Power(void)
-{
- if(PCE_IsCD)
-  memset(ROMSpace + 0x68 * 8192, 0x00, 262144);
-
- if(arcade_card)
-  arcade_card->Power();
-}
-
-MDFNGI EmulatedPCE_Fast =
-{
- PCESettings,
- MDFN_MASTERCLOCK_FIXED(PCE_MASTER_CLOCK),
- 0,
-
- true,                           // Multires possible?
-
- 0,                              // lcm_width
- 0,                              // lcm_height
- NULL,                           // Dummy
-
- MEDNAFEN_CORE_GEOMETRY_BASE_W,  // Nominal width
- MEDNAFEN_CORE_GEOMETRY_BASE_H,  // Nominal height
-
- FB_WIDTH,                       // Framebuffer width
- FB_HEIGHT,                      // Framebuffer height
-
- 2,                              // Number of output sound channels
-};
 
 static bool ReadM3U(std::vector<std::string> &file_list, std::string path, unsigned depth = 0)
 {
@@ -1068,13 +197,7 @@ MDFNGI *MDFNI_LoadGame(const char *force_module, const char *name)
 	// End load per-game settings
 	//
 
-   if(Load(name, GameFile) <= 0)
-   {
-      file_close(GameFile);
-      MDFN_indent(-2);
-      MDFNGameInfo = NULL;
-      return(0);
-   }
+   Load(name, GameFile);
 
 	MDFN_LoadGameCheats(NULL);
 	MDFNMP_InstallReadPatches();
