+++ mednafen/mednafen.h
@@ -18,14 +18,9 @@
 #define GET_FDATA(fp) (fp.f_data)
 #define GET_FSIZE(fp) (fp.f_size)
 #define GET_FEXTS(fp) (fp.f_ext)
-#define GET_FDATA_PTR(fp) (fp->f_data)
-#define GET_FSIZE_PTR(fp) (fp->f_size)
-#define GET_FEXTS_PTR(fp) (fp->f_ext)
-#define gzopen(a, b) fopen(a, b)
-#define gzread(a, b, c) fread(b, c, 1, a)
-#define gzclose(a) fclose(a)
-#define gzgetc(a) fgetc(a)
-#define gzseek(a,b,c) fseek(a,b,c)
+#define GET_FDATA_PTR(fp) (fp->data)
+#define GET_FSIZE_PTR(fp) (fp->size)
+#define GET_FEXTS_PTR(fp) (fp->ext)
 
 extern MDFNGI *MDFNGameInfo;
