+++ mednafen/pce_fast/pce.h
@@ -1,11 +1,27 @@
 #ifndef _PCE_H
 #define _PCE_H
 
+#include <vector>
+
 #include "../mednafen-types.h"
 #include "../mednafen.h"
 #include "../state.h"
 #include "../general.h"
 #include "../mednafen-memory.h"
+#include "psg.h"
+
+#define MEDNAFEN_CORE_NAME_MODULE "pce_fast"
+#define MEDNAFEN_CORE_NAME "Mednafen SuperGrafx"
+#define MEDNAFEN_CORE_VERSION "v0.9.41"
+#define MEDNAFEN_CORE_EXTENSIONS "pce|sgx|cue|ccd|chd"
+#define MEDNAFEN_CORE_TIMING_FPS 59.82
+#define MEDNAFEN_CORE_GEOMETRY_BASE_W 512
+#define MEDNAFEN_CORE_GEOMETRY_BASE_H 243
+#define MEDNAFEN_CORE_GEOMETRY_MAX_W 512
+#define MEDNAFEN_CORE_GEOMETRY_MAX_H 243
+#define MEDNAFEN_CORE_GEOMETRY_ASPECT_RATIO (6.0 / 5.0)
+#define FB_WIDTH 512
+#define FB_HEIGHT 243
 
 #define PCE_MASTER_CLOCK        21477272.727273
 
@@ -32,5 +48,17 @@ extern int pce_overclocked;
 
 extern uint8 BaseRAM[32768 + 8192];
 extern bool PCE_IsCD;
+
+int LoadCD(std::vector<CDIF *> *CDInterfaces);
+void Load(const char *name, MDFNFILE *fp);
+void DoSimpleCommand(int cmd);
+void CloseGame(void);
+void Emulate(EmulateSpecStruct *espec);
+
+extern PCEFast_PSG *psg;
+extern bool IsSGX;
+
+extern uint8 ROMSpace[0x88 * 8192 + 8192];
+extern uint8 BaseRAM[32768 + 8192];
 
 #endif
